requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.2
django-betterforms==2.0.0
django-bootstrap5==23.1
django-clearcache==1.2.1
django-debug-toolbar==4.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
